pyproject.toml
[tool.poetry]
name = "discord-gpt-bot"
version = "0.1.0"
description = ""
authors = ["ekusiadadus <ekusiadadus@gmail.com>"]
readme = "README.md"
packages = [{include = "discord_gpt_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.2.2"
whisper = "^1.1.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
